d9ioao

+++ xchacha20/my_utils.c
@@ -213,6 +213,6 @@ void *memcpy(void *__restrict __dest, const void *__restrict __src, size_t __n)
 {
    char *csrc = (char *)__src;
    char *cdest = (char *)__dest;
-   for (int i=0; i<__n; i++)
+   for (size_t i=0; i<__n; i++)
        cdest[i] = csrc[i];
 }
